refactor: Step1 학습 테스트 코드 패키지 이동
