+++ src/test/java/study/step1/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.step1;
 
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
